requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.8
django-redis==6.0.0
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
kombu==5.5.4
numpy==2.3.4
packaging==25.0
pandas==2.3.3
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
redis==7.0.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
supervisor==4.3.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
